Move form/transaction management to FormAndTransactionProvider
