+++ packages/composite-checkout/src/components/checkout-provider.tsx
@@ -3,27 +3,18 @@ import { useI18n } from '@wordpress/react-i18n';
 import debugFactory from 'debug';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import CheckoutContext from '../lib/checkout-context';
-import { useFormStatusManager } from '../lib/form-status';
 import { LineItemsProvider } from '../lib/line-items';
 import defaultTheme from '../lib/theme';
-import { useTransactionStatusManager } from '../lib/transaction-status';
 import {
 	validateArg,
 	validateLineItems,
 	validatePaymentMethods,
 	validateTotal,
 } from '../lib/validation';
-import { LineItem, CheckoutProviderProps, FormStatus, TransactionStatus } from '../types';
+import { LineItem, CheckoutProviderProps } from '../types';
 import CheckoutErrorBoundary from './checkout-error-boundary';
-import { FormStatusProvider } from './form-status-provider';
-import TransactionStatusHandler from './transaction-status-handler';
-import { TransactionStatusProvider } from './transaction-status-provider';
-import type {
-	PaymentEventCallback,
-	PaymentErrorCallback,
-	PaymentProcessorResponseData,
-	CheckoutContextInterface,
-} from '../types';
+import { FormAndTransactionProvider } from './form-and-transaction-provider';
+import type { CheckoutContextInterface } from '../types';
 
 const debug = debugFactory( 'composite-checkout:checkout-provider' );
 
@@ -92,25 +83,6 @@ export function CheckoutProvider( {
 		setPaymentMethodId
 	);
 
-	// Keep track of form status state (loading, validating, ready, etc).
-	const formStatusManager = useFormStatusManager( Boolean( isLoading ), Boolean( isValidating ) );
-
-	// Keep track of transaction status state (pending, complete, redirecting, etc).
-	const transactionStatusManager = useTransactionStatusManager();
-	const { transactionLastResponse, transactionStatus, transactionError } = transactionStatusManager;
-
-	// When form status or transaction status changes, call the appropriate callbacks.
-	useCallEventCallbacks( {
-		onPaymentComplete,
-		onPaymentRedirect,
-		onPaymentError,
-		formStatus: formStatusManager.formStatus,
-		transactionError,
-		transactionStatus,
-		paymentMethodId,
-		transactionLastResponse,
-	} );
-
 	// Create a big blob of state to store in React Context for use by all this Provider's children.
 	const value: CheckoutContextInterface = useMemo(
 		() => ( {
@@ -146,14 +118,16 @@ export function CheckoutProvider( {
 			<CheckoutProviderPropValidator propsToValidate={ propsToValidate } />
 			<ThemeProvider theme={ theme || defaultTheme }>
 				<LineItemsProvider items={ items } total={ total }>
-					<TransactionStatusProvider transactionStatusManager={ transactionStatusManager }>
-						<FormStatusProvider formStatusManager={ formStatusManager }>
-							<CheckoutContext.Provider value={ value }>
-								<TransactionStatusHandler redirectToUrl={ redirectToUrl } />
-								{ children }
-							</CheckoutContext.Provider>
-						</FormStatusProvider>
-					</TransactionStatusProvider>
+					<FormAndTransactionProvider
+						onPaymentComplete={ onPaymentComplete }
+						onPaymentRedirect={ onPaymentRedirect }
+						onPaymentError={ onPaymentError }
+						isLoading={ isLoading }
+						isValidating={ isValidating }
+						redirectToUrl={ redirectToUrl }
+					>
+						<CheckoutContext.Provider value={ value }>{ children }</CheckoutContext.Provider>
+					</FormAndTransactionProvider>
 				</LineItemsProvider>
 			</ThemeProvider>
 		</CheckoutErrorBoundary>
@@ -183,72 +157,6 @@ function CheckoutProviderPropValidator( {
 		validatePaymentMethods( paymentMethods );
 	}, [ items, paymentMethods, paymentProcessors, propsToValidate, total ] );
 	return null;
-}
-
-// When form status or transaction status changes, call the appropriate callbacks.
-function useCallEventCallbacks( {
-	onPaymentComplete,
-	onPaymentRedirect,
-	onPaymentError,
-	formStatus,
-	transactionError,
-	transactionStatus,
-	paymentMethodId,
-	transactionLastResponse,
-}: {
-	onPaymentComplete?: PaymentEventCallback;
-	onPaymentRedirect?: PaymentEventCallback;
-	onPaymentError?: PaymentErrorCallback;
-	formStatus: FormStatus;
-	transactionError: string | null;
-	transactionStatus: TransactionStatus;
-	paymentMethodId: string | null;
-	transactionLastResponse: PaymentProcessorResponseData;
-} ): void {
-	// Store the callbacks as refs so we do not call them more than once if they
-	// are anonymous functions. This way they are only called when the
-	// transactionStatus/formStatus changes, which is what we really want.
-	const paymentCompleteRef = useRef( onPaymentComplete );
-	paymentCompleteRef.current = onPaymentComplete;
-	const paymentRedirectRef = useRef( onPaymentRedirect );
-	paymentRedirectRef.current = onPaymentRedirect;
-	const paymentErrorRef = useRef( onPaymentError );
-	paymentErrorRef.current = onPaymentError;
-
-	const prevFormStatus = useRef< FormStatus >();
-	const prevTransactionStatus = useRef< TransactionStatus >();
-
-	useEffect( () => {
-		if (
-			paymentCompleteRef.current &&
-			formStatus === FormStatus.COMPLETE &&
-			formStatus !== prevFormStatus.current
-		) {
-			debug( "form status changed to complete so I'm calling onPaymentComplete" );
-			paymentCompleteRef.current( { paymentMethodId, transactionLastResponse } );
-		}
-		prevFormStatus.current = formStatus;
-	}, [ formStatus, transactionLastResponse, paymentMethodId ] );
-
-	useEffect( () => {
-		if (
-			paymentRedirectRef.current &&
-			transactionStatus === TransactionStatus.REDIRECTING &&
-			transactionStatus !== prevTransactionStatus.current
-		) {
-			debug( "transaction status changed to redirecting so I'm calling onPaymentRedirect" );
-			paymentRedirectRef.current( { paymentMethodId, transactionLastResponse } );
-		}
-		if (
-			paymentErrorRef.current &&
-			transactionStatus === TransactionStatus.ERROR &&
-			transactionStatus !== prevTransactionStatus.current
-		) {
-			debug( "transaction status changed to error so I'm calling onPaymentError" );
-			paymentErrorRef.current( { paymentMethodId, transactionError } );
-		}
-		prevTransactionStatus.current = transactionStatus;
-	}, [ transactionStatus, paymentMethodId, transactionLastResponse, transactionError ] );
 }
 
 // Reset the selected payment method if the list of payment methods changes.
